# further latency reductions