--- a/src/callbacks_step/glm_speed.jl
+++ b/src/callbacks_step/glm_speed.jl
@@ -15,14 +15,14 @@ struct GlmSpeedCallback{RealT<:Real}
 end
 
 
-function Base.show(io::IO, cb::DiscreteCallback{Condition,Affect!}) where {Condition, Affect!<:GlmSpeedCallback}
+function Base.show(io::IO, @nospecialize cb::DiscreteCallback{<:Any, <:GlmSpeedCallback})
   glm_speed_callback = cb.affect!
   @unpack glm_scale, cfl = glm_speed_callback
   print(io, "GlmSpeedCallback(glm_scale=", glm_scale, ", cfl=", cfl, ")")
 end
 
 
-function Base.show(io::IO, ::MIME"text/plain", cb::DiscreteCallback{Condition,Affect!}) where {Condition, Affect!<:GlmSpeedCallback}
+function Base.show(io::IO, ::MIME"text/plain", @nospecialize cb::DiscreteCallback{<:Any, <:GlmSpeedCallback})
   if get(io, :compact, false)
     show(io, cb)
   else