--- a/src/callbacks_step/lbm_collision.jl
+++ b/src/callbacks_step/lbm_collision.jl
@@ -15,12 +15,12 @@ end
 lbm_collision_callback(u, t, integrator) = !isfinished(integrator)
 
 
-function Base.show(io::IO, cb::DiscreteCallback{Condition,Affect!}) where {Condition, Affect!<:typeof(lbm_collision_callback)}
+function Base.show(io::IO, @nospecialize cb::DiscreteCallback{<:Any, <:typeof(lbm_collision_callback)})
   print(io, "LBMCollisionCallback()")
 end
 
 
-function Base.show(io::IO, ::MIME"text/plain", cb::DiscreteCallback{Condition,Affect!}) where {Condition, Affect!<:typeof(lbm_collision_callback)}
+function Base.show(io::IO, ::MIME"text/plain", @nospecialize cb::DiscreteCallback{<:Any, <:typeof(lbm_collision_callback)})
   if get(io, :compact, false)
     show(io, cb)
   else