Additionally, this tests whether renaming `master` to `main` worked as expected.