@ranocha 2021-02-11
Good suggestion! I added some explanation to the docs.

Here are some timings I get from running 
```bash
julia16 --threads=1 --check-bounds=no -e '@time using Trixi; @time trixi_include("examples/2d/elixir_advection_amr.jl"); @time trixi_include("examples/2d/elixir_euler_ec.jl"); @time trixi_include("examples/2d/elixir_mhd_blast_wave.jl", tspan=(0.0, 1.0e-3))'
```
three times.

|               |    `main`    |    This PR   |
|:-------------:|:------------:|:------------:|
| `using Trixi` |  3.4 ± 0.3 s |  3.4 ± 0.3 s |
|   Advection   | 29.2 ± 0.3 s | 28.6 ± 0.1 s |
|     Euler     |  6.6 ± 0.1 s |  6.4 ± 0.2 s |
|      MHD      | 13.4 ± 0.1 s | 13.1 ± 0.1 s |