@sloede 2021-02-11
Not trying to be unpopular, but I am wondering if this optimization is really worth it. As far as I can tell, it's at best a ~3% improvement in startup time, while the code noise grows considerably with these `@nospecialize` scattered around. My initial feeling is that effort-vs-impact is not really that good... What do you think, @trixi-framework/principal-developers?