@ranocha 2021-02-12
I don't insist on merging this as it is. Just another data point: Precompiling Trixi on `main` takes 23.8 s while this PR reduces that to 20.5 s on my system (measured using `julia16 --threads=1 --check-bounds=no -e '@time using Trixi'`). Since some of you complained about the time it takes to precompile Trixi after modifying something, you might want to consider this, too.

Alternatively, we can also just delete the `@nospecialize` statements. Nevertheless, some of the additional/changed precompile statement should still be considered (since the output format changed after they were added).