@sloede 2021-02-12
Hm. Yeah, I also noticed the slowdown in precompilation. The above improvement sounds better, but I am still not sure if it's worth it. I'll try to get a hold of the others today.